Refactor: Moves login repositories and data source to its layers : dat and domain
