--- app/src/main/java/com/jabama/challenge/repository/oauth/AccessTokenDataSourceImpl.kt
@@ -1,8 +1,0 @@
-package com.jabama.challenge.repository.oauth
-
-import com.jabama.challenge.network.oauth.AccessTokenService
-import com.jabama.challenge.network.oauth.RequestAccessToken
-
-class AccessTokenDataSourceImpl(private val accessTokenService: AccessTokenService) : AccessTokenDataSource {
-    override fun accessToken(requestAccessToken: RequestAccessToken) = accessTokenService.accessToken(requestAccessToken)
-}
