+++ login/src/main/java/com/example/login/domain/TokenRepository.kt
@@ -1,4 +1,4 @@
-package com.jabama.challenge.repository.token
+package com.example.login.domain
 
 import kotlinx.coroutines.Deferred
